src/main/java/com/lucianaugusto/recipeapp/services/ImageServiceImpl.java
package com.lucianaugusto.recipeapp.services;

import java.io.IOException;

import org.springframework.stereotype.Service;
import org.springframework.web.multipart.MultipartFile;

import com.lucianaugusto.recipeapp.domain.Recipe;
import com.lucianaugusto.recipeapp.repositories.RecipeRepository;

import lombok.extern.slf4j.Slf4j;

@Slf4j
@Service
public class ImageServiceImpl implements ImageService {

	private final RecipeRepository recipeRepository;

	public ImageServiceImpl(RecipeRepository recipeRepository) {
		this.recipeRepository = recipeRepository;
	}

	@Override
	public void saveImageFile(String recipeId, MultipartFile file) {
		log.debug("Received a file");
		try {
			Recipe recipe = recipeRepository.findById(recipeId).get();

			Byte[] byteObjects = new Byte[file.getBytes().length];

			int i = 0;

			for (byte b : file.getBytes()) { // Since there's not autoboxing for arrays, we have to do it manually from
												// scratch
				byteObjects[i++] = b;
			}

			recipe.setImage(byteObjects);

			recipeRepository.save(recipe);

		} catch (IOException e) {
			// TODO improve error handling
			log.error("Error occurred " + e);
			e.printStackTrace();
		}

	}
}
